chore: remove set galleries feature toggle
Set-driven galleries are now always enabled, and Contentful-driven
galleries deprecated.
