+++ packages/portal/.env.example
@@ -127,9 +127,6 @@ OAUTH_CLIENT="YOUR_CLIENT"
 # Enable entity management
 # ENABLE_ENTITY_MANAGEMENT=0
 
-# Enable set driven galleries
-# ENABLE_SET_GALLERIES=0
-
 # URL of the Europeana oEmbed endpoint
 # EUROPEANA_OEMBED_PROVIDER_URL=https://oembed.europeana.eu
