+++ packages/portal/src/features/toggles.js
@@ -5,7 +5,6 @@ export default [
   { name: 'fieldBoosting' },
   { name: 'jiraServiceDeskFeedbackForm' },
   { name: 'rejectEntityRecommendations' },
-  { name: 'setGalleries' },
   { name: 'storiesPageAllTags' },
   { name: 'translatedItems' },
   { name: 'galleryPublicationSubmissions' }
