+++ packages/portal/src/pages/galleries/_.vue
@@ -21,47 +21,8 @@
       :illustration-src="$fetchState.error.illustrationSrc"
       :status-code="$fetchState.error.statusCode"
     />
-    <b-container
-      v-else-if="!setGalleriesEnabled"
-    >
-      <ContentWarningModal
-        v-if="contentWarning"
-        :title="contentWarning.name"
-        :description="contentWarning.description"
-        :page-slug="`galleries/${identifier}`"
-      />
-      <ContentHeader
-        :title="pageMeta.title"
-        :description="htmlDescription"
-        :media-url="shareMediaUrl"
-        :context-label="$tc('galleries.galleries', 1)"
-      />
-      <b-row class="flex-md-row pb-5">
-        <b-col cols="12">
-          <div
-            v-masonry
-            transition-duration="0"
-            item-selector=".card"
-            horizontal-order="true"
-            column-width=".masonry-container .card"
-            class="masonry-container"
-            data-qa="gallery images"
-          >
-            <ContentCard
-              v-for="image in images"
-              :key="image.identifier"
-              v-masonry-tile
-              :title="imageTitle(image)"
-              :image-url="imageUrl(image)"
-              :lazy="false"
-              :url="{ name: 'item-all', params: { pathMatch: image.identifier.slice(1) } }"
-            />
-          </div>
-        </b-col>
-      </b-row>
-    </b-container>
     <div
-      v-else-if="set.id"
+      v-if="set.id"
       class="mt-n3"
       data-qa="user gallery page"
     >
@@ -243,11 +204,6 @@
   import redirectToPrefPathMixin from '@/mixins/redirectToPrefPath';
   import pageMetaMixin from '@/mixins/pageMeta';
 
-  // TODO: the following imports are only needed for contentful Galleries.
-  import ContentHeader from '../../components/generic/ContentHeader';
-  import { marked } from 'marked';
-  import stripMarkdown from '@/mixins/stripMarkdown';
-
   export default {
     name: 'GalleryPage',
     components: {
@@ -259,27 +215,19 @@
       SocialShareModal,
       SetFormModal: () => import('@/components/set/SetFormModal'),
       SetRecommendations: () => import('@/components/set/SetRecommendations'),
-      // TODO: The following components are only used in contentful galleries
-      ContentHeader,
-      ContentCard: () => import('../../components/generic/ContentCard'),
-      ContentWarningModal: () => import('@/components/generic/ContentWarningModal'),
       PublishSetButton: () => import('@/components/set/PublishSetButton'),
       SetPublicationRequestWidget: () => import('@/components/set/SetPublicationRequestWidget')
 
     },
     mixins: [
       redirectToPrefPathMixin,
-      pageMetaMixin,
-      // TODO: markdown is only used in contentful galleries
-      stripMarkdown
+      pageMetaMixin
     ],
     async beforeRouteLeave(_to, _from, next) {
-      if (this.setGalleriesEnabled) {
-        await this.$store.commit('set/setActive', null);
-        await this.$store.commit('set/setActiveRecommendations', []);
-        await this.$store.commit('entity/setFeaturedSetId', null);
-        await this.$store.commit('entity/setPinned', []);
-      }
+      await this.$store.commit('set/setActive', null);
+      await this.$store.commit('set/setActiveRecommendations', []);
+      await this.$store.commit('entity/setFeaturedSetId', null);
+      await this.$store.commit('entity/setPinned', []);
       next();
     },
     data() {
@@ -292,20 +240,16 @@
       };
     },
     async fetch() {
-      if (this.setGalleriesEnabled) {
-        try {
-          this.validateRoute();
-          await this.$store.dispatch('set/fetchActive', this.setId);
-          this.redirectToPrefPath('galleries-all', this.setId, this.set.title.en);
-          if (this.setIsEntityBestItems && this.userIsEntityEditor) {
-            await this.$store.commit('entity/setFeaturedSetId', this.setId);
-            await this.$store.dispatch('entity/getPins');
-          }
-        } catch (error) {
-          this.handleFetchError(error);
+      try {
+        this.validateRoute();
+        await this.$store.dispatch('set/fetchActive', this.setId);
+        this.redirectToPrefPath('galleries-all', this.setId, this.set.title.en);
+        if (this.setIsEntityBestItems && this.userIsEntityEditor) {
+          await this.$store.commit('entity/setFeaturedSetId', this.setId);
+          await this.$store.dispatch('entity/getPins');
         }
-      } else {
-        await this.fetchContentfulGallery();
+      } catch (error) {
+        this.handleFetchError(error);
       }
     },
     computed: {
@@ -316,9 +260,6 @@
           ogType: 'article',
           ogImage: this.shareMediaUrl
         };
-      },
-      setGalleriesEnabled() {
-        return this.$features.setGalleries;
       },
       set() {
         return this.$store.state.set.active || {};
@@ -376,35 +317,16 @@
         return this.$tc(label, this.set.total, { max });
       },
       displayTitle() {
-        // TODO: remove contentful gallery fallback
-        if (this.setGalleriesEnabled) {
-          return langMapValueForLocale(this.set.title, this.$i18n.locale);
-        }
-        return { values: [this.title] };
+        return langMapValueForLocale(this.set.title, this.$i18n.locale);
       },
       displayDescription() {
-        // TODO: remove contentful gallery fallback
-        if (this.setGalleriesEnabled) {
-          return langMapValueForLocale(this.set.description, this.$i18n.locale);
-        }
-        return { values: [this.description] };
+        return langMapValueForLocale(this.set.description, this.$i18n.locale);
       },
       shareMediaUrl() {
-        // TODO: remove contentful gallery fallback
-        if (this.setGalleriesEnabled) {
-          return this.$apis.thumbnail.edmPreview(this.set?.items?.[0]?.edmPreview?.[0], { size: 400 });
-        }
-        return this.images.length === 0 ? null : this.imageUrl(this.images[0]);
-      },
-      // TODO: description & htmlDescription are only used for contentful galleries.
-      description() {
-        return this.stripMarkdown(this.rawDescription);
-      },
-      htmlDescription() {
-        return marked.parse(this.rawDescription);
+        return this.$apis.thumbnail.edmPreview(this.set?.items?.[0]?.edmPreview?.[0], { size: 400 });
       },
       displayMetadata() {
-        return this.set.visibility === ('private' || 'published') || this.set.creator.nickname;
+        return this.set.visibility === ('private' || 'published') || this.set.creator?.nickname;
       },
       weaveUrl() {
         return `https://experience.weave-culture.eu/import/europeana/set/${this.setId}`;
@@ -435,40 +357,6 @@
           },
           params: { profile: 'standard' }
         });
-      },
-      async fetchContentfulGallery() {
-        // TODO: Remove this method, when using set driven galleries.
-        const variables = {
-          identifier: this.$route.params.pathMatch,
-          locale: this.$i18n.isoLocale(),
-          preview: this.$route.query.mode === 'preview'
-        };
-
-        await this.$contentful.query('galleryPage', variables)
-          .then(response => response.data.data)
-          .then(data => {
-            if (data.imageGalleryCollection.items.length === 0) {
-              throw createHttpError(404, this.$t('messages.notFound'));
-            }
-
-            const gallery = data.imageGalleryCollection.items[0];
-
-            this.contentWarning = gallery.contentWarning;
-            this.identifier = variables.identifier;
-            this.images = gallery.hasPartCollection.items.filter(image => image !== null);
-            this.rawDescription = gallery.description;
-            this.title = gallery.name;
-          });
-      },
-      imageTitle(data) {
-        if (data.encoding) {
-          return data.encoding.dcTitleLangAware || data.encoding.dcDescriptionLangAware || this.$t('record.record');
-        }
-        return data.name;
-      },
-      imageUrl(data) {
-        const edmPreview = data.encoding?.edmPreview?.[0] || data.thumbnailUrl;
-        return this.$apis.thumbnail.edmPreview(edmPreview, { size: 400 });
       },
       handleFetchError(error) {
         if (process.server) {
