--- packages/portal/src/pages/set/_.vue
@@ -1,373 +1,0 @@
-<template>
-  <div
-    :class="$fetchState.error && 'white-page'"
-  >
-    <b-container
-      v-if="$fetchState.pending"
-      data-qa="loading spinner container"
-    >
-      <b-row class="flex-md-row py-4 text-center">
-        <b-col cols="12">
-          <LoadingSpinner />
-        </b-col>
-      </b-row>
-    </b-container>
-    <ErrorMessage
-      v-else-if="$fetchState.error"
-      data-qa="error message container"
-      :error="$fetchState.error.message"
-      :title-path="$fetchState.error.titlePath"
-      :description-path="$fetchState.error.descriptionPath"
-      :illustration-src="$fetchState.error.illustrationSrc"
-      :status-code="$fetchState.error.statusCode"
-    />
-    <div
-      v-else-if="set.id"
-      class="mt-n3"
-      data-qa="user gallery page"
-    >
-      <b-container
-        fluid
-      >
-        <b-row class="flex-md-row pt-5 bg-white mb-4">
-          <b-col
-            cols="12"
-          >
-            <b-container class="mb-5">
-              <b-row class="mb-4">
-                <b-col>
-                  <div
-                    class="context-label"
-                  >
-                    {{ $tc('galleries.galleries', 1) }}
-                  </div>
-                  <h1
-                    :lang="displayTitle.code"
-                  >
-                    {{ displayTitle.values[0] }}
-                  </h1>
-                  <p
-                    class="usergallery-description mb-3 w-75"
-                    :lang="displayDescription.code"
-                  >
-                    {{ displayDescription.values[0] }}
-                  </p>
-                  <!-- TODO: to avoid showing an empty div + whitespace, the v-if is on the div
-                      This can be changed when this functionality is further developed
-                  -->
-                  <div
-                    v-if="set.visibility === 'private' || set.creator.nickname"
-                    class="usergallery-metadata mb-2"
-                  >
-                    <span
-                      v-if="set.creator.nickname"
-                      class="curator mb-2"
-                    >
-                      {{ $t('set.labels.curatedBy') }} @{{ set.creator.nickname }}
-                    </span>
-                    <span
-                      v-if="set.visibility === 'private'"
-                      class="
-                      visibility mb-2"
-                    >
-                      {{ $t('set.labels.private') }}
-                    </span>
-                  </div>
-                </b-col>
-              </b-row>
-              <div class="d-inline-flex collection-buttons">
-                <template
-                  v-if="userIsOwner"
-                >
-                  <b-button
-                    variant="outline-primary"
-                    class="text-decoration-none mr-2"
-                    data-qa="edit set button"
-                    @click="$bvModal.show(setFormModalId)"
-                  >
-                    {{ $t('actions.edit') }}
-                  </b-button>
-                  <SetFormModal
-                    :set-id="set.id"
-                    :modal-id="setFormModalId"
-                    :title="set.title"
-                    :description="set.description"
-                    :visibility="set.visibility"
-                  />
-                </template>
-                <template v-if="set.visibility !== 'private'">
-                  <b-button
-                    v-b-modal.share-modal
-                    variant="outline-primary"
-                    class="text-decoration-none"
-                  >
-                    {{ $t('actions.share') }}
-                  </b-button>
-                  <SocialShareModal
-                    :media-url="shareMediaUrl"
-                    :share-to="[{
-                      identifier: 'weavex',
-                      name: 'WEAVEx',
-                      url: weaveUrl,
-                      tooltip: $t('set.shareTo.weavex.tooltip')
-                    }]"
-                  />
-                </template>
-              </div>
-            </b-container>
-          </b-col>
-        </b-row>
-      </b-container>
-      <b-container
-        class="mb-3"
-        data-qa="user set"
-      >
-        <b-row>
-          <b-col>
-            <h2
-              class="related-heading text-uppercase"
-              data-qa="item count"
-            >
-              {{ displayItemCount }}
-            </h2>
-          </b-col>
-        </b-row>
-        <b-row>
-          <b-col>
-            <b-container class="px-0">
-              <b-row class="mb-3">
-                <b-col cols="12">
-                  <ItemPreviewCardGroup
-                    :items="set.items"
-                    :show-pins="setIsEntityBestItems && userIsEntityEditor"
-                    :draggable-items="userIsOwner"
-                    @endItemDrag="reorderItems"
-                  />
-                </b-col>
-              </b-row>
-            </b-container>
-          </b-col>
-        </b-row>
-        <client-only>
-          <SetRecommendations
-            v-if="displayRecommendations"
-            :identifier="`/${$route.params.pathMatch}`"
-            :type="set.type"
-          />
-        </client-only>
-      </b-container>
-    </div>
-  </div>
-</template>
-
-<script>
-  // TODO: This file will be deprecated when set-driven galleries are in production.
-  // TODO: Also move the beforeRouteEnter redirect to the legacy middleware.
-  import ClientOnly from 'vue-client-only';
-  import createHttpError from 'http-errors';
-  import pageMetaMixin from '@/mixins/pageMeta';
-
-  import {
-    ITEM_URL_PREFIX as EUROPEANA_DATA_URL_ITEM_PREFIX,
-    SET_URL_PREFIX as EUROPEANA_DATA_URL_SET_PREFIX
-  } from '@/plugins/europeana/data';
-  import { langMapValueForLocale } from  '@/plugins/europeana/utils';
-
-  import ItemPreviewCardGroup from '@/components/item/ItemPreviewCardGroup';
-  import SocialShareModal from '@/components/sharing/SocialShareModal.vue';
-
-  export default {
-    name: 'SetPage',
-
-    components: { ClientOnly,
-                  ErrorMessage: () => import('@/components/generic/ErrorMessage'),
-                  LoadingSpinner: () => import('@/components/generic/LoadingSpinner'),
-                  ItemPreviewCardGroup,
-                  SocialShareModal,
-                  SetFormModal: () => import('@/components/set/SetFormModal'),
-                  SetRecommendations: () => import('@/components/set/SetRecommendations') },
-
-    mixins: [pageMetaMixin],
-
-    async beforeRouteLeave(to, from, next) {
-      await this.$store.commit('set/setActive', null);
-      await this.$store.commit('set/setActiveRecommendations', []);
-      await this.$store.commit('entity/setFeaturedSetId', null);
-      await this.$store.commit('entity/setPinned', []);
-      next();
-    },
-
-    middleware({ app, params, redirect }) {
-      if (app.$features.setGalleries) {
-        redirect({ name: `galleries-all___${app.i18n.locale}`, params });
-      }
-    },
-
-    async fetch() {
-      try {
-        this.validateRoute();
-
-        await this.$store.dispatch('set/fetchActive', this.setId);
-
-        if (this.setIsEntityBestItems && this.userIsEntityEditor) {
-          await this.$store.commit('entity/setFeaturedSetId', this.setId);
-          await this.$store.dispatch('entity/getPins');
-        }
-      } catch (error) {
-        if (process.server) {
-          this.$nuxt.context.res.statusCode = error.statusCode || 500;
-        }
-        if (error.statusCode === 403) {
-          error.titlePath = 'errorMessage.galleryUnauthorised.title';
-          error.descriptionPath = 'errorMessage.galleryUnauthorised.description';
-          error.pageTitlePath = 'errorMessage.galleryUnauthorised.metaTitle';
-          error.illustrationSrc = require('@/assets/img/illustrations/il-gallery-unauthorised.svg');
-        }
-        throw error;
-      }
-    },
-
-    computed: {
-      pageMeta() {
-        return {
-          title: this.displayTitle.values[0],
-          description: this.displayDescription?.values?.[0],
-          ogType: 'article',
-          ogImage: this.shareMediaUrl
-        };
-      },
-      fetchState() {
-        return this.$fetchState;
-      },
-      set() {
-        return this.$store.state.set.active || {};
-      },
-      setId() {
-        return this.$route.params.pathMatch;
-      },
-      setFormModalId() {
-        return `set-form-modal-${this.setId}`;
-      },
-      setCreatorId() {
-        return this.set.creator && typeof this.set.creator === 'string' ? this.set.creator : this.set.creator.id;
-      },
-      userIsOwner() {
-        return this.$auth.loggedIn && this.$store.state.auth.user &&
-          this.setCreatorId &&
-          this.setCreatorId.endsWith(`/${this.$store.state.auth.user.sub}`);
-      },
-      userIsEntityEditor() {
-        return this.$auth.userHasClientRole('entities', 'editor') &&
-          this.$auth.userHasClientRole('usersets', 'editor');
-      },
-      userCanEdit() {
-        return this.userIsOwner || (this.setIsEntityBestItems && this.userIsEntityEditor);
-      },
-      setIsEntityBestItems() {
-        return this.set.type === 'EntityBestItemsSet';
-      },
-      displayTitle() {
-        return langMapValueForLocale(this.set.title, this.$i18n.locale);
-      },
-      displayDescription() {
-        return langMapValueForLocale(this.set.description, this.$i18n.locale);
-      },
-      displayRecommendations() {
-        return this.enableRecommendations && this.$auth.loggedIn && this.userCanEdit;
-      },
-      enableRecommendations() {
-        if (this.setIsEntityBestItems) {
-          return this.$features.acceptEntityRecommendations ||
-            this.$features.rejectEntityRecommendations;
-        }
-        return true;
-      },
-      displayItemCount() {
-        const max = 100;
-        const label = this.set.total > max ? 'items.itemOf' : 'items.itemCount';
-        return this.$tc(label, this.set.total, { max });
-      },
-      shareMediaUrl() {
-        return this.$apis.thumbnail.edmPreview(this.set?.items?.[0]?.edmPreview?.[0], { size: 400 });
-      },
-      weaveUrl() {
-        return `https://experience.weave-culture.eu/import/europeana/set/${this.setId}`;
-      }
-    },
-
-    methods: {
-      validateRoute() {
-        if (!/^\d+$/.test(this.$route.params.pathMatch)) {
-          throw createHttpError(400, 'Invalid set ID');
-        }
-      },
-
-      reorderItems(items) {
-        this.$store.dispatch('set/update', {
-          id: `${EUROPEANA_DATA_URL_SET_PREFIX}/${this.setId}`,
-          body: {
-            type: this.set.type,
-            title: this.set.title,
-            description: this.set.description,
-            visibility: this.set.visibility,
-            items: items.map(item => `${EUROPEANA_DATA_URL_ITEM_PREFIX}${item.id}`)
-          },
-          params: { profile: 'standard' }
-        });
-      }
-    }
-  };
-</script>
-
-<style lang="scss" scoped>
-  @import '@/assets/scss/variables';
-  @import '@/assets/scss/icons';
-
-  .usergallery-description {
-    color: $mediumgrey;
-  }
-
-  .usergallery-metadata {
-    font-size: $font-size-small;
-    font-weight: 600;
-    line-height: 1.125;
-    color: $mediumgrey;
-
-    .curator,
-    .visibility {
-      display: inline-flex;
-      align-items: center;
-
-      &::before {
-        font-size: 1.5rem;
-        padding-right: 0.2rem;
-      }
-    }
-
-    .curator {
-      margin-right: 1.5rem;
-
-      &::before {
-        @extend %icon-font;
-
-        content: '\e92e';
-        font-size: 1.125rem;
-      }
-    }
-
-    .visibility {
-      &::before {
-        @extend %icon-font;
-
-        content: '\e92d';
-        font-size: 1.125rem;
-      }
-    }
-  }
-
-  .collection-buttons {
-    .text {
-      font-weight: 600;
-    }
-  }
-</style>
