+++ tests/e2e/features/pages/gallery-page.feature
@@ -1,16 +1,14 @@
 Feature: Gallery page
 
   Scenario: View a gallery
-
     When I open the `gallery page`
-    Then I see a `page title`
-    And I see a `content card` in the `gallery images`
+    And I see an `item preview` in the `item previews grid`
     And I see a `card title` in the `card body`
     And I am on an accessible page
     And I should have a Europeana branded page title
 
   Scenario: Go to a item
     When I open the `gallery page`
-    Then I see a `content card` in the `gallery images`
-    And I click a `content card`
+    And I see an `item preview` in the `item previews grid`
+    And I click an `item preview`
     Then I see an `item page`
